+++ CHANGELOG.md
@@ -33,6 +33,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Internally added an `EphemeralSegment` to aid with parsing efficiency
   without altering the end structure of the query.
 
+### Removed
+- Removed `BaseSegment.grammar`, `BaseSegment._match_grammar()` and
+  `BaseSegment._parse_grammar()` instead preferring references directly
+  to `BaseSegment.match_grammar` and `BaseSegment.parse_grammar`.
+
 ## [0.3.6] - 2020-09-24
 
 ### Added